requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2019.6.16
cffi==1.14.5
channels==3.0.4
channels-redis==3.2.0
chardet==3.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
future==0.18.2
futures==3.1.1
goslate==1.5.1
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
lxml==4.6.3
MouseInfo==0.1.3
msgpack==1.0.2
oauthlib==3.1.0
pep517==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycountry==20.7.3
pycparser==2.20
PyDictionary==2.0.1
PyGetWindow==0.0.9
PyJWT==1.7.1
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pyperclip==1.8.2
Pyphen==0.10.0
PyRect==0.1.4
PyScreeze==0.1.27
python-bidi==0.4.2
python3-openid==3.1.0
pytube==10.8.2
pytube3==9.6.4
PyTweening==1.0.3
pytz==2019.2
pywhatkit==5.1
reportlab==3.5.67
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.0
tinycss2==1.1.0
toml==0.10.2
twilio==6.60.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.25.3
validate-email==1.3
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
wikipedia==1.4.0
xhtml2pdf==0.2.5
xlwt==1.3.0
zope.interface==5.4.0
